Fix : PlayerGetDamage damage값 int로 넘겨서 HP에 전달

+++ Assets/Scripts/Ingame/Entities/PlayerCharacter.cs
@@ -137,7 +137,7 @@ namespace Cardevil.Ingame.Entities
         public void PlayerGetDamage(float amount)
         {
             Debug.Log($"Player takes {amount} damage!");
-            Managers.Game.PlayerStatus.CurrentHp -= amount;
+            Managers.Game.PlayerStatus.CurrentHp -= (int)amount;
         }
 
         public async UniTask TurnMove()
